backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3
email-validator==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
Pillow==10.1.0

# Web scraping and scheduling
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
APScheduler==3.10.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1